SSM-137: removed santizer
